requirements.txt
uvicorn
fastapi
sqlalchemy
pytest
httpx
python-jose
python-multipart
argon2-cffi
pydantic[email]
python-dotenv